+++ src-tauri/gen/android/app/src/main/java/com/folio/folioreader/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.folio.folioapp
+package com.folio.folioreader
 
 import android.Manifest
 import android.app.SearchManager